requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
Brotli==1.0.9
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
discord==2.3.2
discord.py==2.3.2
frozenlist==1.4.0
idna==3.4
multidict==6.0.4
mutagen==1.46.0
pycparser==2.21
pycryptodomex==3.18.0
PyNaCl==1.5.0
python-dotenv==1.0.0
websockets==11.0.3
yarl==1.9.2
yt-dlp==2023.7.6